backend/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
markup==0.2
Pillow==9.4.0
python-dotenv==0.21.0
pytz==2022.7
sqlparse==0.4.3
